@@ -14,7 +14,7 @@ description: A Flutter E-commerce App.
 version: 0.0.2+2
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -31,14 +31,20 @@ dependencies:
   equatable: ^1.1.0
   flutter_translate: any
   bubble_tab_indicator: any
-  dartz: ^0.8.9
+  http: ^0.12.0+4
   rxdart: ^0.23.1
   sqflite:
   path:
-
+  flutter_secure_storage: ^3.3.1+1
+  font_awesome_flutter: 8.7.0
+  # check internet connection availability
+  data_connection_checker: ^0.3.4
+  get_it: 
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  mockito: ^4.1.1
 
 dependency_overrides:
   pedantic: ^1.9.0
@@ -130,30 +136,3 @@ flutter:
         - asset: fonts/Metropolis-BlackItalic.otf
           weight: 900
           style: italic
-
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
