@@ -5,86 +5,102 @@
 import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:openflutterecommerce/data/abstract/category_repository.dart';
-import 'package:openflutterecommerce/data/abstract/model/category.dart';
-import 'package:openflutterecommerce/data/abstract/product_repository.dart';
+import 'package:openflutterecommerce/data/abstract/favorites_repository.dart';
+import 'package:openflutterecommerce/data/abstract/model/product.dart';
+import 'package:openflutterecommerce/data/abstract/model/sort_rules.dart';
 import 'package:openflutterecommerce/data/fake_model/hashtag_repository.dart';
+import 'package:openflutterecommerce/domain/usecases/products/find_products_by_filter_use_case.dart';
+import 'package:openflutterecommerce/domain/usecases/products/products_by_filter_params.dart';
+import 'package:openflutterecommerce/domain/usecases/products/products_by_filter_result.dart';
+import 'package:openflutterecommerce/locator.dart';
 import 'package:openflutterecommerce/presentation/features/products/products.dart';
-import 'package:openflutterecommerce/presentation/widgets/widgets.dart';
+import 'bloc_list_data.dart';
 
-class ProductsBloc extends Bloc<ProductsListEvent, ProductState> {
-  final ProductRepository productRepository;
+class ProductsBloc extends Bloc<ProductsListEvent, ProductsState> {
+  final FindProductsByFilterUseCase findProductsByFilterUseCase;
   final CategoryRepository categoryRepository;
+  final FavoritesRepository favoritesRepository;
   final HashtagRepository hashtagRepository;
   final int categoryId;
 
   ProductsBloc({
     @required this.categoryId,
     @required this.categoryRepository,
-    @required this.productRepository,
+    @required this.favoritesRepository,
     @required this.hashtagRepository,
-  }) : assert(productRepository != null);
+  }) : findProductsByFilterUseCase = sl();
 
   @override
-  ProductState get initialState => ProductInitialState();
+  ProductsState get initialState => ProductsListViewState();
 
   @override
-  Stream<ProductState> mapEventToState(ProductsListEvent event) async* {
+  Stream<ProductsState> mapEventToState(ProductsListEvent event) async* {
     if (event is ScreenLoadedEvent) {
-      final data = await getStateData(categoryId);
-      yield ProductsLoadedState(
-          isLoading: false,
-          showSortBy: false,
-          sortBy: SortBy.Popular,
-          selectedBrandIds: <int>[],
-          selectedCategories: <Category>[],
-          selectedSizes: <String>[],
-          selectedColors: <Color>[],
-          priceRange: RangeValues(70, 150),
-          availablePriceRange: RangeValues(0, 1000),
-          brandSearchKey: '',
-          data: data);
-    } else if (event is ProductChangeSortByEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(sortBy: event.sortBy, showSortBy: false);
-    } else if (event is ProductShowSortByEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(showSortBy: true);
-    } else if (event is ProductChangePriceRangeEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(priceRange: event.priceRange);
-    } else if (event is ProductChangeSelectedColorsEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(selectedColors: event.selectedColors);
-    } else if (event is ProductChangeSelectedSizesEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(selectedSizes: event.selectedSizes);
-    } else if (event is ProductChangeSelectedCategoriesEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(selectedCategories: event.selectedCategories);
-    } else if (event is ProductChangeSelectedBrandsEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(selectedBrandIds: event.selectedBrandIds);
-    } else if (event is ProductChangeBrandSearchKeyEvent) {
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(brandSearchKey: event.newSearchKey);
-    } else if (event is MakeFavorite) {
-      if (event.isFavorite) {
-        await productRepository.addToFavorites(event.productId);
+      ProductListData data = await getInitialStateData(categoryId);
+      yield ProductsListViewState(
+          sortBy: SortRules(),
+          data: data,
+          filterRules: data.filterRules);
+    } else if (event is ProductsChangeViewEvent) {
+      if (state is ProductsListViewState) {
+        yield (state as ProductsListViewState).getTiles();
       } else {
-        await productRepository.removeFromFavorites(event.productId);
+        yield (state as ProductsTileViewState).getList();
       }
-      final state = this.state as ProductsLoadedState;
-      yield state.copyWith(data: await getStateData(categoryId));
+    } else if (event is ProductChangeSortRulesEvent) {
+      yield state.getLoading();
+      ProductsByFilterResult productResults = await findProductsByFilterUseCase.execute(
+        ProductsByFilterParams(
+          categoryId: categoryId,
+          filterRules: state.filterRules,
+          sortBy: event.sortBy
+        ));
+      final List<Product> filteredData = productResults.products;
+      yield state.copyWith(
+        sortBy: event.sortBy,
+        data: state.data.copyWith(filteredData),
+      );
+    } else if (event is ProductChangeFilterRulesEvent) {
+      yield state.getLoading();
+      ProductsByFilterResult productResults = await findProductsByFilterUseCase.execute(
+        ProductsByFilterParams(
+          categoryId: categoryId,
+          filterRules: event.filterRules,
+          sortBy: state.sortBy
+        ));
+      final List<Product> filteredData = productResults.products;
+      yield state.copyWith(
+          filterRules: event.filterRules,
+          data: state.data.copyWith(filteredData));
+    } else if (event is MakeFavoriteEvent) {
+      if (event.isFavorite) {
+        await favoritesRepository.addToFavorites(
+            event.productId, event.favoriteAttributes);
+      } else {
+        await favoritesRepository.removeFromFavorites(event.productId);
+      }
+      final List<Product> data = state.data.products;
+      yield state.copyWith(
+          data: state.data.copyWith(data.map((item) {
+        if (event.productId == item.id) {
+          return item.favorite(event.isFavorite);
+        } else {
+          return item;
+        }
+      }).toList(growable: false)));
     }
   }
 
-  Future<ProductStateData> getStateData(int categoryId) async {
-    var data = ProductStateData();
-
-    data.products = await productRepository.getProductsInCategory(categoryId);
-    data.hashtags = hashtagRepository.getHashtags();
-    data.category = await categoryRepository.getCategoryDetails(categoryId);
-
-    return data;
+  Future<ProductListData> getInitialStateData(int categoryId) async {
+    ProductsByFilterResult productResults = await findProductsByFilterUseCase.execute(
+      ProductsByFilterParams(
+        categoryId: categoryId
+      )
+    );
+    return ProductListData(
+        productResults.products,
+        hashtagRepository.getHashtags(),
+        await categoryRepository.getCategoryDetails(categoryId),
+        productResults.filterRules);
   }
 }
