@@ -1,12 +1,14 @@
+import 'package:openflutterecommerce/data/abstract/model/product_attribute.dart';
 import 'package:openflutterecommerce/domain/entities/entity.dart';
 
 class ProductEntity extends Entity<int> {
   final String title;
-  final String image;
+  final String subTitle;
+  final List<String> images;
   final String thumb;
   final double price;
-  final double discountPercent;
-  final int categoryId;
+  final int discountPercent;
+  final List<int> categoryIds;
   final int amount;
   final String description;
   final bool isFavourite;
@@ -16,35 +18,51 @@ class ProductEntity extends Entity<int> {
   final int rating3Count;
   final int rating4Count;
   final int rating5Count;
+  final List<ProductAttribute> selectableAttributes;
 
   ProductEntity(
     {int id,
     this.title,
-    this.image,
+    this.subTitle,
+    this.images,
     this.thumb,
-    this.price,
-    this.discountPercent,
-    this.categoryId,
+    price,
+    discountPercent,
+    categoryIds,
     this.amount,
     this.description,
-    this.isFavourite,
-    this.rating,
-    this.rating1Count,
-    this.rating2Count,
-    this.rating3Count,
-    this.rating4Count,
-    this.rating5Count}) : super(id);
+    this.selectableAttributes,
+    bool isFavourite,
+    rating,
+    rating1Count,
+    rating2Count,
+    rating3Count,
+    rating4Count,
+    rating5Count}) :
+      rating = rating??0,
+      rating1Count = rating1Count??0,
+      rating2Count = rating2Count??0,
+      rating3Count = rating3Count??0,
+      rating4Count = rating4Count??0,
+      rating5Count = rating5Count??0,
+      isFavourite = isFavourite??false,
+      discountPercent = discountPercent??0,
+      price = (price??0).toDouble(),
+      categoryIds = categoryIds??[],
+      super(id);
 
   @override
   Map<String, dynamic> toMap() {
     return {
       'id': id,
       'title': title,
-      'image': image,
+      //TODO: serialize all images and add tests
+      'image': images.isNotEmpty?images[0]:'',
       'thumb': thumb,
       'price': price,
       'discountPercent': discountPercent,
-      'categoryId': categoryId,
+      //TODO: serialize all categoryIds and add tests
+      'categoryId': categoryIds.isNotEmpty?categoryIds[0]:0,
       'amount': amount,
       'description': description,
       'isFavourite': isFavourite,
