@@ -6,25 +6,37 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_translate/flutter_translate.dart';
 import 'package:openflutterecommerce/config/routes.dart';
 import 'package:openflutterecommerce/config/theme.dart';
+import 'package:openflutterecommerce/data/abstract/favorites_repository.dart';
+import 'package:openflutterecommerce/data/abstract/model/filter_rules.dart';
 import 'package:openflutterecommerce/data/abstract/product_repository.dart';
+import 'package:openflutterecommerce/data/abstract/user_repository.dart';
 import 'package:openflutterecommerce/data/fake_model/fake_product_repository.dart';
+import 'package:openflutterecommerce/data/fake_model/fake_user_repository.dart';
+import 'package:openflutterecommerce/presentation/features/forget_password/forget_password_screen.dart';
+import 'package:openflutterecommerce/presentation/features/sign_in/sign_in.dart';
+import 'package:openflutterecommerce/presentation/features/filters/filters_screen.dart';
 import 'package:openflutterecommerce/presentation/features/product_details/product_screen.dart';
 import 'package:openflutterecommerce/presentation/features/products/products.dart';
+import 'package:openflutterecommerce/presentation/features/sign_in/signin_screen.dart';
+import 'package:openflutterecommerce/presentation/features/sign_up/signup_screen.dart';
 import 'package:openflutterecommerce/presentation/features/splash_screen.dart';
 
 import 'config/routes.dart';
+import 'data/abstract/cart_repository.dart';
 import 'data/abstract/category_repository.dart';
+import 'data/fake_model/fake_cart_repository.dart';
 import 'data/fake_model/fake_category_repository.dart';
 import 'presentation/features/authentication/authentication.dart';
+import 'presentation/features/forget_password/forget_password.dart';
+import 'presentation/features/sign_up/sign_up_bloc.dart';
 import 'presentation/features/cart/cart.dart';
 import 'presentation/features/categories/categories.dart';
 import 'presentation/features/checkout/checkout.dart';
 import 'presentation/features/favorites/favorites.dart';
 import 'presentation/features/home/home.dart';
 import 'presentation/features/profile/profile.dart';
-import 'presentation/features/signin/forget_password.dart';
-import 'presentation/features/signin/signin.dart';
-import 'presentation/features/signin/signup.dart';
+
+import 'locator.dart' as service_locator;
 
 class SimpleBlocDelegate extends BlocDelegate {
   @override
@@ -47,6 +59,8 @@ class SimpleBlocDelegate extends BlocDelegate {
 }
 
 void main() async {
+  await service_locator.init();
+
   var delegate = await LocalizationDelegate.create(
     fallbackLocale: 'en_US',
     supportedLocales: ['en_US', 'de', 'fr'],
@@ -69,6 +83,15 @@ void main() async {
           ),
           RepositoryProvider<ProductRepository>(
             create: (context) => FakeProductRepository(),
+          ),
+          RepositoryProvider<FavoritesRepository>(
+            create: (context) => FakeProductRepository(),
+          ),
+          RepositoryProvider<UserRepository>(
+            create: (context) => FakeUserRepository(),
+          ),
+          RepositoryProvider<CartRepository>(
+            create: (context) => FakeCartRepository(),
           ),
         ],
         child: LocalizedApp(
@@ -129,21 +152,29 @@ class OpenFlutterEcommerceApp extends StatelessWidget {
 
   BlocProvider<ForgetPasswordBloc> _buildForgetPasswordBloc() {
     return BlocProvider<ForgetPasswordBloc>(
-      create: (context) => ForgetPasswordBloc(),
+      create: (context) => ForgetPasswordBloc(
+        userRepository: RepositoryProvider.of<UserRepository>(context),
+      ),
       child: ForgetPasswordScreen(),
     );
   }
 
   BlocProvider<SignInBloc> _buildSignInBloc() {
     return BlocProvider<SignInBloc>(
-      create: (context) => SignInBloc(),
+      create: (context) => SignInBloc(
+        userRepository: RepositoryProvider.of<UserRepository>(context),
+        authenticationBloc: BlocProvider.of<AuthenticationBloc>(context),
+      ),
       child: SignInScreen(),
     );
   }
 
   BlocProvider<SignUpBloc> _buildSignUpBloc() {
     return BlocProvider<SignUpBloc>(
-      create: (context) => SignUpBloc(),
+      create: (context) => SignUpBloc(
+        userRepository: RepositoryProvider.of<UserRepository>(context),
+        authenticationBloc: BlocProvider.of<AuthenticationBloc>(context),
+      ),
       child: SignUpScreen(),
     );
   }
@@ -170,6 +201,11 @@ class OpenFlutterEcommerceApp extends StatelessWidget {
       final ProductDetailsParameters parameters = settings.arguments;
       return MaterialPageRoute(builder: (context) {
         return ProductDetailsScreen(parameters);
+      });
+    } else if (settings.name == OpenFlutterEcommerceRoutes.filters) {
+      final FilterRules filterRules = settings.arguments;
+      return MaterialPageRoute(builder: (context) {
+        return FiltersScreen(filterRules);
       });
     } else {
       return MaterialPageRoute(
