@@ -2,17 +2,19 @@
 // Author: openflutterproject@gmail.com
 // Date: 2020-02-06
 import 'package:bloc/bloc.dart';
-import 'package:flutter/material.dart';
-import 'package:openflutterecommerce/data/abstract/category_repository.dart';
+import 'package:openflutterecommerce/data/abstract/model/category.dart';
+import 'package:openflutterecommerce/domain/usecases/categories/categories_by_filter_params.dart';
+import 'package:openflutterecommerce/domain/usecases/categories/find_categories_by_filter_use_case.dart';
+import 'package:openflutterecommerce/locator.dart';
 
 import 'categories.dart';
 
 class CategoryBloc extends Bloc<CategoryEvent, CategoryState> {
-  final CategoryRepository _categoryRepository;
+  final FindCategoriesByFilterUseCase findCategoriesByFilterUseCase;
 
-  CategoryBloc({
-    @required CategoryRepository categoryRepository,
-  }) : _categoryRepository = categoryRepository;
+  CategoryBloc(): findCategoriesByFilterUseCase = sl();
+
+
 
   @override
   CategoryState get initialState => CategoryLoadingState();
@@ -23,15 +25,13 @@ class CategoryBloc extends Bloc<CategoryEvent, CategoryState> {
       if (state is CategoryListViewState) {
         if (state.parentCategoryId != event.parentCategoryId) {
           yield CategoryLoadingState();
-          final categories = await _categoryRepository.getCategories(
-              parentCategoryId: event.parentCategoryId);
+          List<Category> categories = await _getCategoriesByFilter(event.parentCategoryId);
           yield CategoryListViewState(
               categories: categories, parentCategoryId: event.parentCategoryId);
         }
       } else {
         yield CategoryLoadingState();
-        final categories = await _categoryRepository.getCategories(
-            parentCategoryId: event.parentCategoryId);
+        List<Category> categories = await _getCategoriesByFilter(event.parentCategoryId);
         yield CategoryListViewState(
             parentCategoryId: event.parentCategoryId, categories: categories);
       }
@@ -39,22 +39,19 @@ class CategoryBloc extends Bloc<CategoryEvent, CategoryState> {
       if (state is CategoryTileViewState) {
         if (state.parentCategoryId != event.parentCategoryId) {
           yield CategoryLoadingState();
-          final categories = await _categoryRepository.getCategories(
-              parentCategoryId: event.parentCategoryId);
+          List<Category> categories = await _getCategoriesByFilter(event.parentCategoryId);
           yield CategoryTileViewState(
               categories: categories, parentCategoryId: event.parentCategoryId);
         }
       } else {
         yield CategoryLoadingState();
-        final categories = await _categoryRepository.getCategories(
-            parentCategoryId: event.parentCategoryId);
+        List<Category> categories = await _getCategoriesByFilter(event.parentCategoryId);
         yield CategoryTileViewState(
-            parentCategoryId: event.parentCategoryId, categories: categories);
+          parentCategoryId: event.parentCategoryId, categories: categories);
       }
     } else if (event is ChangeCategoryParent) {
       yield CategoryLoadingState();
-      final categories = await _categoryRepository.getCategories(
-          parentCategoryId: event.parentCategoryId);
+      List<Category> categories = await _getCategoriesByFilter(event.parentCategoryId);
       if (state is CategoryTileViewState) {
         yield CategoryTileViewState(
             parentCategoryId: event.parentCategoryId, categories: categories);
@@ -63,5 +60,13 @@ class CategoryBloc extends Bloc<CategoryEvent, CategoryState> {
             parentCategoryId: event.parentCategoryId, categories: categories);
       }
     }
+  }
+  Future<List<Category>> _getCategoriesByFilter(int categoryId) async{
+    final categoriesData = await findCategoriesByFilterUseCase.execute(
+      CategoriesByFilterParams(
+        categoryId: categoryId
+      )
+    );
+    return categoriesData.categories;
   }
 }
