@@ -4,119 +4,54 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:openflutterecommerce/config/routes.dart';
 import 'package:openflutterecommerce/config/theme.dart';
-import 'package:openflutterecommerce/data/abstract/model/product.dart';
-import 'package:openflutterecommerce/presentation/widgets/independent/scaffold_collapsing.dart';
-import 'package:openflutterecommerce/presentation/widgets/widgets.dart';
+import 'package:openflutterecommerce/presentation/features/product_details/product_screen.dart';
+import 'package:openflutterecommerce/presentation/widgets/extensions/product_view.dart';
 
-import '../../wrapper.dart';
 import '../favorites_bloc.dart';
 import '../favorites_event.dart';
 import '../favorites_state.dart';
 
-class FavouritesTileView extends StatefulWidget {
-  final Function({@required ViewChangeType changeType, int index}) changeView;
-
-  final FavouriteState state;
-  const FavouritesTileView({Key key, this.changeView, this.state})
-      : super(key: key);
-
-  @override
-  _FavouritesTileViewState createState() => _FavouritesTileViewState();
-}
-
-class _FavouritesTileViewState extends State<FavouritesTileView> {
-  _FavouritesTileViewState({Key key});
-
+class FavouritesTileView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return _buildFavouritesHeader(context, widget.state);
-  }
-
-  Widget _buildFavouritesHeader(BuildContext context, FavouriteState state) {
-    final bloc = BlocProvider.of<FavouriteBloc>(context);
-    var width = MediaQuery.of(context).size.width;
-    var productView = ProductView.CardView;
-    var sortBy = SortBy.Popular;
-
-    return OpenFlutterCollapsingScaffold(
-      background: AppColors.background,
-      title: 'Favourites',
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: <Widget>[
-          Container(
-            color: AppColors.white,
-            child: Column(
-              children: <Widget>[
-                Container(
-                    width: width,
-                    padding: EdgeInsets.only(left: 8.0, right: 8.0),
-                    child: OpenFlutterHashTagList(
-                        tags: state is FavouriteTileViewState
-                            ? state.hashtags
-                            : [],
-                        height: 30)),
-                Container(
-                  padding: EdgeInsets.only(
-                      top: AppSizes.sidePadding * 2.50,
-                      bottom: AppSizes.sidePadding),
-                  width: width * 0.95,
-                  child: OpenFlutterProductFilter(
-                    width: width * 0.95,
-                    height: 20,
-                    productView: productView,
-                    sortBy: sortBy,
-                    onFilterClicked: (() => {print('Filter Clicked')}),
-                    onChangeViewClicked: (() {
-                      print('Show ListView');
-                      bloc..add(ShowListViewEvent());
-                      widget.changeView(changeType: ViewChangeType.Backward);
-                    }),
-                    onSortClicked: ((SortBy sortBy) => {print('Sort Clicked')}),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          Expanded(
-            child: GridView.extent(
-                childAspectRatio: 1 / 1.8,
-                maxCrossAxisExtent: width / 2,
-                padding: const EdgeInsets.all(4),
-                mainAxisSpacing: 4,
-                crossAxisSpacing: 4,
-                shrinkWrap: true,
-                children: buildProductList(state, width)),
-          )
-        ],
-      ),
-      bottomMenuIndex: 3,
-    );
-  }
-
-  List<Widget> buildProductList(FavouriteState state, double width) {
-    var elements = <Widget>[];
-    var widgetWidth = (width) / 2;
-    var height = widgetWidth * 1.789;
-    var products =
-        state is FavouriteTileViewState ? state.favouriteProducts : <Product>[];
-    for (var i = 0; i < products.length; i++) {
-      elements.add(
-        OpenFlutterProductCard(
-          product: products[i],
-          height: height,
-          width: widgetWidth,
-          showCartButton: true,
-          showRemoveButton: true,
-          showColorAndSize: true,
-          showRatingInLine: true,
-          showTopLabel: true,
-          showCategoryFirst: true,
+    return BlocBuilder<FavouriteBloc, FavouriteState>(
+        builder: (context, state) {
+      return SliverGrid(
+        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+            crossAxisCount: 2,
+            mainAxisSpacing: 4.0,
+            crossAxisSpacing: 4.0,
+            childAspectRatio: 1.589),
+        delegate: SliverChildBuilderDelegate(
+          (BuildContext context, int index) {
+            return Padding(
+              padding: EdgeInsets.symmetric(horizontal: AppSizes.sidePadding),
+              child: state.data[index].getTileView(
+                context: context,
+                showProductInfo: () {
+                  Navigator.of(context).pushNamed(
+                      OpenFlutterEcommerceRoutes.product,
+                      arguments: ProductDetailsParameters(
+                          state.data[index].product.id,
+                          state.data[index].product.categoryIds.isNotEmpty ? 
+                            state.data[index].product.categoryIds[0]: 0,
+                          selectedAttributes: state.data[index].favoriteForm));
+                },
+                onRemoveFromFavorites: () {
+                  BlocProvider.of<FavouriteBloc>(context).add(
+                      RemoveFromFavoriteEvent(state.data[index].product.id));
+                },
+                onAddToCart: () {
+                  BlocProvider.of<FavouriteBloc>(context)
+                      .add(AddToCartEvent(state.data[index].product.id));
+                },
+              ),
+            );
+          },
         ),
       );
-    }
-    return elements;
+    });
   }
 }
