@@ -1,3 +1,5 @@
+import 'dart:collection';
+
 import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -5,15 +7,27 @@ import 'package:openflutterecommerce/data/abstract/model/product_attribute.dart'
 
 @immutable
 class ProductEvent extends Equatable {
+  
   @override
   List<Object> get props => [];
 }
 
 @immutable
-class ScreenLoadedEvent extends ProductEvent {}
+class ProductScreenLoadedEvent extends ProductEvent {
+  final int productId;
+  final int categoryId;
+
+  ProductScreenLoadedEvent({
+    @required this.productId, 
+    @required this.categoryId
+  });}
 
 @immutable
-class ProductAddToFavoritesEvent extends ProductEvent {}
+class ProductAddToFavoritesEvent extends ProductEvent {
+  final HashMap<ProductAttribute, String> selectedAttributes;
+
+  ProductAddToFavoritesEvent(this.selectedAttributes);
+}
 
 @immutable
 class ProductRemoveFromFavoritesEvent extends ProductEvent {}
