@@ -1,3 +1,5 @@
+import 'dart:collection';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:openflutterecommerce/config/theme.dart';
@@ -27,20 +29,20 @@ class ProductDetailsView extends StatefulWidget {
       @required this.similarProducts,
       this.category,
       this.hasReviews = false})
-      : super(key: key);
+      : assert(product!=null),
+       super(key: key);
 
   @override
   _ProductDetailsViewState createState() => _ProductDetailsViewState();
 }
 
 class _ProductDetailsViewState extends State<ProductDetailsView> {
-  List<int> list = [1, 2, 3, 4, 5];
   Orientation orientation;
   bool favorite;
 
   @override
   void initState() {
-    favorite = widget.product.isFavorite;
+    favorite = widget.product?.isFavorite??false;
     super.initState();
   }
 
@@ -79,13 +81,14 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
                         Container(
                           height: deviceHeight * 0.52,
                           child: ListView.builder(
-                            itemBuilder: (context, index) => Image.asset(
-                              'assets/images/products/shortdress.png',
-                              width: deviceWidth * 0.75,
-                              height: deviceHeight * 0.52,
+                            itemBuilder: (context, index) => Padding(
+                              padding: EdgeInsets.only(right: AppSizes.sidePadding),
+                              child: Image.network(
+                                state.product.images[index].address
+                              )
                             ),
                             scrollDirection: Axis.horizontal,
-                            itemCount: list.length,
+                            itemCount: state.product.images.length,
                           ),
                         ),
                         Container(
@@ -94,18 +97,23 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
                           child: Row(
                               mainAxisAlignment: MainAxisAlignment.spaceBetween,
                               mainAxisSize: MainAxisSize.max,
-                              children: state.product.attributes
+                              children: 
+                                 (state.product.selectableAttributes != null ?
+                                   state.product.selectableAttributes
                                       .map((value) => selectionOutlineButton(
                                           deviceWidth,
                                           value,
-                                          state.selectedAttributes[value]))
-                                      .toList() +
+                                          state.selectedAttributes == null
+                                              ? null
+                                              : state
+                                                  .selectedAttributes[value]))
+                                      .toList() : List<Widget>()) +
                                   [
                                     OpenFlutterFavouriteButton(
                                       favourite: favorite,
                                       setFavourite: () => {setFavourite(bloc)},
                                     )
-                                  ]),
+                                  ]) ,
                         ),
                         productDetails(_theme),
                         //Function call for Product detail widget
@@ -184,7 +192,8 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
 
   void setFavourite(ProductBloc bloc) {
     if (!favorite) {
-      bloc.add(ProductAddToFavoritesEvent());
+      bloc.add(ProductAddToFavoritesEvent(
+          HashMap())); //TODO ask for real parameters if required
     } else {
       bloc.add(ProductRemoveFromFavoritesEvent());
     }
@@ -299,10 +308,11 @@ class _ProductDetailsViewState extends State<ProductDetailsView> {
   }
 
   void _addItemToCart(BuildContext context, ProductLoadedState state) async {
-    if (state.selectedAttributes.length == state.product.attributes.length) {
+    if (state.selectedAttributes.length ==
+        state.product.selectableAttributes.length) {
       BlocProvider.of<ProductBloc>(context).add(ProductAddToCartEvent());
     } else {
-      for (final attribute in state.product.attributes) {
+      for (final attribute in state.product.selectableAttributes) {
         if (!state.selectedAttributes.containsKey(attribute)) {
           await _showSelectAttributeBottomSheet(context, attribute);
         }
