@@ -2,7 +2,6 @@ export 'data_driven/box_value_select.dart';
 export 'data_driven/category_list_element.dart';
 export 'data_driven/category_tile.dart';
 export 'data_driven/hashtag_list.dart';
-export 'data_driven/product_card.dart';
 export 'data_driven/product_list_view.dart';
 export 'data_driven/rating_summary.dart';
 export 'independent/block_header.dart';
@@ -19,7 +18,6 @@ export 'independent/input_button.dart';
 export 'independent/input_field.dart';
 export 'independent/label_right_checkbox.dart';
 export 'independent/price_slider.dart';
-export 'independent/product_filter.dart';
 export 'independent/product_rating.dart';
 export 'independent/product_review_item.dart';
 export 'independent/right_arrow_action.dart';
@@ -27,4 +25,4 @@ export 'independent/scaffold.dart';
 export 'independent/service_button.dart';
 export 'independent/sortby.dart';
 export 'independent/text_tile.dart';
-export 'product_tile.dart';
+export 'independent/view_options.dart';
