@@ -18,12 +18,12 @@ void main() {
       ProductEntity product = ProductEntity(
         id: 1,
         title: 'Product 1',
-        image: 'https:example.com/image.png',
+        images: ['https:example.com/image.png'],
         thumb: 'https://example.com/thumbnail.png',
         price: 289.5,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -43,12 +43,12 @@ void main() {
       ProductEntity product = ProductEntity(
         id: 1,
         title: 'Product 1',
-        image: 'https:example.com/image.png',
+        images: ['https:example.com/image.png'],
         thumb: 'https://example.com/thumbnail.png',
         price: 289.5,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -65,12 +65,12 @@ void main() {
       ProductEntity productToUpdate = ProductEntity(
         id: insertedProduct.id,
         title: 'Product 2',
-        image: 'https:example.com/image.png',
+        images: ['https:example.com/image.png'],
         thumb: 'https://example.com/thumbnail.png',
         price: 500,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -92,12 +92,12 @@ void main() {
       await productDataSource.insert(ProductEntity(
         id: 1,
         title: 'Product 1',
-        image: 'https:example.com/image.png',
+        images: ['https:example.com/image.png'],
         thumb: 'https://example.com/thumbnail.png',
         price: 289.5,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -110,12 +110,12 @@ void main() {
       await productDataSource.insert(ProductEntity(
         id: 2,
         title: 'Product 2',
-        image: 'https:example.com/image2.png',
+        images: ['https:example.com/image2.png'],
         thumb: 'https://example.com/thumbnail2.png',
         price: 289.5,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -128,12 +128,12 @@ void main() {
       await productDataSource.insert(ProductEntity(
         id: 3,
         title: 'Product 3',
-        image: 'https:example.com/image3.png',
+        images: ['https:example.com/image3.png'],
         thumb: 'https://example.com/thumbnail3.png',
         price: 700,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -156,12 +156,12 @@ void main() {
       await productDataSource.insert(ProductEntity(
         id: 1,
         title: 'Product 1',
-        image: 'https:example.com/image.png',
+        images: ['https:example.com/image.png'],
         thumb: 'https://example.com/thumbnail.png',
         price: 289.5,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -184,12 +184,12 @@ void main() {
       await productDataSource.insert(ProductEntity(
         id: 1,
         title: 'Product 1',
-        image: 'https:example.com/image.png',
+        images: ['https:example.com/image.png'],
         thumb: 'https://example.com/thumbnail.png',
         price: 289.5,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -202,12 +202,12 @@ void main() {
       await productDataSource.insert(ProductEntity(
         id: 2,
         title: 'Product 2',
-        image: 'https:example.com/image2.png',
+        images: ['https:example.com/image2.png'],
         thumb: 'https://example.com/thumbnail2.png',
         price: 289.5,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
@@ -220,12 +220,12 @@ void main() {
       await productDataSource.insert(ProductEntity(
         id: 3,
         title: 'Product 3',
-        image: 'https:example.com/image3.png',
+        images: ['https:example.com/image3.png'],
         thumb: 'https://example.com/thumbnail3.png',
         price: 700,
         discountPercent: 5.0,
         isFavourite: true,
-        categoryId: 1,
+        categoryIds: [1],
         amount: 255,
         description: 'Best product in the market',
         rating: 4.2,
